pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4","pydantic>=2.6.4","numpy>=1.26.4","py-jsonic>=0.0.2"]
build-backend = "flit_core.buildapi"

[project]
name = "CompoST"
authors = [{name = "Jakub Kucera"}]
description = "Set of prescribed objects used to store and manipulate composites related desing/manufacture/lifecycle information"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
version = "0.7.5"
dependencies = ["pydantic>=2.6.4","numpy>=1.26.4","py-jsonic>=0.0.2"]

[project.urls]
Home = "https://github.com/National-Composites-Centre/CompoST"
